Initial commit - extract 16 colors from theme, apply them to Linux
Extracts the dominant colours from a subset of built-in faces, and applies them
to the Linux session with wal.

The method used to apply the colors is currently an external python script that
modifies the pywal theme. This script will need to be integrated internally
before distribution. The Python dependency is fine because Pywal itself requires
Python.

Does not currently theme GTK apps.