Protect against nil colors & failed color form evaluation
If a color form fails to evaluate, catch the error and throw nil. Do not
interrupt the program. Ignore these nil values wherever possible and try to
evaluate other color forms instead.

This is useful when, for example, the user does not have one of the preferred
faces installed. Perhaps a `helm` face is used, but they're running `ivy`. It
also adds a layer of protection against unforeseen color evaluation errors.