Number the colors that are shown to the user when applying
When colors are applied, a list was given before. For the user to visually
inspect the list, they would have to count the colors to determine which color
was being inserted into which index.

Modify the message so it outputs a numbered alist. This should be much easier
for the user to scan - provided they can read elisp alists.