Fix Spacemacs using the default background for left modeline cell
The cell on the far left of the modeline, in Spacemacs, indicates the editing
mode (Emacs vs. evilified - hybrid also seems to be an option). If a theme
doesn't set this color explicitly, it ends up being set to the default color,
(e.g. for Emacs, it's "sky-blue"). This is really an upstream bug with
Spacemacs but it's worth covering for.

The theme should update this face to be in line with the rest of the theme,
otherwise it will clash.